requirements.txt
requests
requests
web3
python-dotenv
fastapi
uvicorn
